requirements.txt
streamlit==1.35.0
streamlit-authenticator==0.3.2
pandas==2.2.2
plotly==5.22.0
PyYAML==6.0.1
bcrypt==4.1.3
